% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_background_pop}
\alias{generate_background_pop}
\title{Generate a consistent background population for synthetic registers}
\usage{
generate_background_pop(
  years_to_generate = 2015:2024,
  pnrs_per_year = 1000,
  turnover_per_year = 20,
  seed = 2109
)
}
\arguments{
\item{years_to_generate}{A vector of years the simulation is intended to cover.
Used to calculate the total pool size needed to simulate population turnover.}

\item{pnrs_per_year}{Integer. The target population size per year. Defaults to 1000.}

\item{turnover_per_year}{Integer. The "buffer" of extra individuals added per year to simulate churn/turnover.
Defaults to 20.}

\item{seed}{Integer. The random seed used to ensure reproducibility of the PNR list. Defaults to 2109.}
}
\value{
A \code{tibble} containing the stable population pool with variables:
\itemize{
\item \code{pnr}: Random 12-digit Personal ID string.
\item \code{opr_land}: Origin country code (ensures at least one of every foreign code).
\item \code{koen}: Gender code.
\item \code{foed_dag}: Birth date.
}
}
\description{
Creates a "master" tibble of unique individuals with stable attributes (PNR, gender, birthday, origin).
This tibble serves as the consistency anchor for other synthetic register functions (e.g., \code{\link{generate_bef}}),
ensuring that the same PNR always corresponds to the same person across different datasets,
and that the synthetic registers are limited to these PNRs in order to facilitate joins.
}
\examples{
# Generate a background population pool
background_pop <- generate_background_pop(
  years_to_generate = 2015:2024,
  pnrs_per_year = 1000,
  turnover_per_year = 20
)
}
